+++ src/vanilla/PluginPy.cs
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
-// 
+//
 
 using AutoRest.Core;
 using AutoRest.Core.Extensibility;
@@ -20,7 +20,7 @@ namespace AutoRest.Python
                 // inherit base settings
                 Context,
 
-                // set code model implementations our own implementations 
+                // set code model implementations our own implementations
                 new Factory<CodeModel, CodeModelPy>(),
                 new Factory<Method, MethodPy>(),
                 new Factory<CompositeType, CompositeTypePy>(),