+++ test/vanilla/Expected/AcceptanceTests/BodyDateTime/bodydatetime/models/__init__.py
@@ -10,9 +10,9 @@
 # --------------------------------------------------------------------------
 
 try:
-    from .error_py3 import Error, ErrorException
+    from ._models_py3 import Error, ErrorException
 except (SyntaxError, ImportError):
-    from .error import Error, ErrorException
+    from ._models import Error, ErrorException
 
 __all__ = [
     'Error', 'ErrorException',